+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v1.3.3...HEAD)
 
+## Added
+- `Nullable`, `Format`, and `Immutable` fields to connector metadata `Property` deserialization.
+
 ## [1.3.3](https://github.com/fivetran/go-fivetran/compare/v1.3.2...v1.3.3)
 
 ## Fixed